requirements.txt
Django==1.6.6
django-plugins
South
python-mimeparse
django-tastypie==0.11.0
email_reply_parser
flufl.bounce
django-markdown-deux
requests
django-extensions
django-haystack
elasticsearch
celery
pyyaml
django-celery
django-autoslug
pytz>=2013b
django-admin-bootstrapped
django-object-actions
unidecode
python-social-auth
django-nose
mock
python-dateutil
#things required in heroku
gunicorn
dj-database-url==0.2.2
dj-static==0.0.5
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
-e git+git@github.com:openpolis/django-popolo.git#egg=django_popolo-dev
